+++ src/double_threshold.cpp
@@ -3,7 +3,7 @@
 #include "canny_stages.h"
 
 void double_threshold(
-    const std::uint8_t input[WIDTH],
+    const std::uint16_t input[WIDTH],
     std::uint8_t output[WIDTH],
     bool valid_in,
     bool *valid_out
@@ -14,7 +14,7 @@ void double_threshold(
     }
 
     for (int column = 0; column < WIDTH; ++column) {
-        const std::uint8_t magnitude = input[column];
+        const std::uint16_t magnitude = input[column];
 
         if (magnitude >= HIGH_THRESHOLD) {
             output[column] = STRONG_EDGE;
